Code Live
ParticleSystem
RotateBehavior
WaypointsBehavior
FireTramp Script

+++ CLASE 19 - PARTICLE SYSTEM/Assets/Scripts/Events/PlayerEvents.cs
@@ -4,10 +4,12 @@ public static class PlayerEvents
 {
     public static event Action OnWin;
     public static event Action<int> OnHeal;
-    public static event Action<int> OnDamage;
+    public static event Action<float> OnDamage;
 
     public static void OnWinCall() { OnWin?.Invoke(); }
     public static void OnHealCall(int hp) { OnHeal?.Invoke(hp); }
-    public static void OnDamageCall(int hp) { OnDamage?.Invoke(hp); }
+    public static void OnDamageCall(float hp) { OnDamage?.Invoke(hp); }
+
+    
 
 }
